pyproject.toml
[project]
name = "fine-tuning-rag"
version = "0.1.0"
description = "Modular Python project with Clean Architecture, DDD and SOLID"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.34.0",
    "typer>=0.15.1",
    "rich>=13.9.4",
    "pydantic>=2.10.3",
    "beautifulsoup4>=4.14.2",
    "openai>=2.7.1",
    "python-dotenv>=1.0.0",
    "requests>=2.32.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.4",
    "mypy>=1.13.0",
]

[project.scripts]
cli = "cli_app.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "cli_app", "http_app"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
files = ["core/**/*.py"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=core --cov-report=term-missing"
